First test on Linux gcc in 2022, first steps:
* INT_shortTIME_emC: cleanup: The definitions of stdint.h should not be placed here, it is only a decision of compl_adaption.h
  The problem was obvious because INT32_MAX was not defined but the types were.
* Base/Time_emC.h definition of INT_shortTIME_emC to use (similar as INT_NUM_emC)
* PIDi_Ctrl_emC.c: feature of disableIntg removed, if necessary can be done with set fI := 0
* PIDi_Ctrl_emC.h: Instead a specific #ifdef DEF_TestVal_PIDi_Ctrl_emC use a common #ifndef DEF_NO_TestValInspc_emC
  It means in normal case all test variable (only for show) are present, especially in test,
  but for a really slow memory processor you can set generally DEF_NO_TestValInspc_emC to save this memory.
* emC_inclComplSpec/cc_TI430_TI/compl_adaption.h: first time newly used since refactoring, now opportune.

+++ emC/Ctrl/T1_Ctrl_emC.h
@@ -87,7 +87,8 @@ static inline int16 dx_T1i_Ctrl_emC(T1i_Ctrl_emC_s* thiz, int16 x) {
 
 
 /**Step routine wrapper for simulink for 16 bit T1 smoothing FBlock. 
- * @simulink Object-FB, no-thizInit, no-thizStep.
+ *                      no-thizInit, 
+ * @simulink Object-FB, no-thizStep.
  */
 static inline void stepY_T1i_Ctrl_emC(T1i_Ctrl_emC_s* thiz, int16 x, int16* y_y) {
   *y_y = step_T1i_Ctrl_emC(thiz, x);
@@ -339,9 +340,9 @@ typedef struct Delayi_Ctrl_emC_T {
 
 
 
-INLINE_emC Delayi_Ctrl_emC_s* ctor_Delayi_Ctrl_emC ( ObjectJc* othiz, int size, int sizeMem ) {
+INLINE_emC Delayi_Ctrl_emC_s* ctor_Delayi_Ctrl_emC ( ObjectJc* othiz, int size, uint sizeMem ) {
 #ifndef __ignoreInCheader_zbnf__
-  ASSERT_emC(sizeMem >= (size-2) * sizeof(int) + sizeof(Delayi_Ctrl_emC_s), "faulty size on Delayi_Ctrl_emC", size, sizeMem); 
+  ASSERT_emC(sizeMem >= (size-2) * (uint)sizeof(int) + (uint)sizeof(Delayi_Ctrl_emC_s), "faulty size on Delayi_Ctrl_emC", size, sizeMem); 
   Delayi_Ctrl_emC_s* thiz = C_CAST(Delayi_Ctrl_emC_s*, othiz);
   thiz->nsize = size;
   return thiz;
